… noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/MirrorReceiverIterator.java
… noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/ProfileBuilder.java
… noisemodelling-pathfinder/src/main/java/org/noise_planet/noisemodelling/pathfinder/utils/ListFunctions.java
… noisemodelling-pathfinder/src/test/java/org/noise_planet/noisemodelling/pathfinder/TestWallReflection.java