Store into a STRTree the visibility cone envelope of a receiver image. Source point envelope is the query to this tree. Performance of this operation if very great:
```
        Before
        
        pointNoiseMap.setMaximumPropagationDistance(800.0);
        pointNoiseMap.setSoundReflectionOrder(1);
        pointNoiseMap.setComputeHorizontalDiffraction(false);
        pointNoiseMap.setComputeVerticalDiffraction(true);        
        Computed in 2464915 ms, 342.30 ms per receiver   
        
        After
        
        pointNoiseMap.setMaximumPropagationDistance(800.0);
        pointNoiseMap.setSoundReflectionOrder(1);
        pointNoiseMap.setComputeHorizontalDiffraction(false);
        pointNoiseMap.setComputeVerticalDiffraction(true);   
        Computed in 1279908 ms, 177.74 ms per receiver
```
Red area is where sound sources will see the receiver reflection
Green square is the envelope used into the STRtree
![image](https://user-images.githubusercontent.com/1382241/154688046-dfdec49b-1f25-4b81-b451-6698ea022f44.png)